+++ CHANGELOG.md
@@ -4,9 +4,9 @@
 - [X] Manage `recycle bin`option for share
 - [X] Manage WSDD2 service
 - [X] Manage Avahi service
+- [X] Veto files for share not global [#79](https://github.com/dianlight/srat/issues/79)
 ###### __🚧 Work in progess__
 - [ ] Manage `local master`option (?)
-- [ ] Veto files for share not global [#79](https://github.com/dianlight/srat/issues/79)
 - [ ] Monitor tab (?)
 - [ ] Help screen or overlay help/tour [#82](https://github.com/dianlight/srat/issues/82)
 - [ ] Custom component [#83](https://github.com/dianlight/srat/issues/83)
@@ -15,8 +15,9 @@
 - [X] `enable`/`disable` share functionality is not working as expected.
 - [X] Renaming the admin user does not correctly create the new user or rename the existing one; issues persist until a full addon reboot.
 ###### __🚧 Work in progess__
+- [ ]  [#80](https://github.com/dianlight/srat/issues/85)
 
- 
+
 ### 🏗 Chore
 - [X] Implemet wachdog
 ###### __🚧 Work in progess__